requirements.txt
anthropic>=0.39
fastapi>=0.115
uvicorn[standard]>=0.30
python-dateutil>=2.9
pyyaml>=6.0
